chore(ci): add building test/ to ci
this commit adds building several applications in the test/ directory to
the project's root level `test.prod` npm script. while wiring these up,
it was determined there are two projects in test/ that do not compile.
fixing these have been deferred to two later stencil tickets, as the
intent of this commit is to get initial building in place